opt: add some app.kubernetes.io/* labels back
