+++ pkg/controllers/tiflash/tasks/pod.go
@@ -145,7 +145,7 @@ func newPod(cluster *v1alpha1.Cluster, tiflash *v1alpha1.TiFlash, configHash str
 			Labels: maputil.Merge(tiflash.Labels, map[string]string{
 				v1alpha1.LabelKeyInstance:   tiflash.Name,
 				v1alpha1.LabelKeyConfigHash: configHash,
-			}),
+			}, k8s.LabelsK8sApp(cluster.Name, v1alpha1.LabelValComponentTiFlash)),
 			Annotations: maputil.Merge(tiflash.GetAnnotations(),
 				k8s.AnnoProm(tiflash.GetMetricsPort(), metricsPath),
 				k8s.AnnoAdditionalProm("tiflash.proxy", tiflash.GetProxyStatusPort())),
