--- pkg/utils/k8s/annotation.go
@@ -1,35 +1,0 @@
-// Copyright 2024 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package k8s
-
-import "fmt"
-
-// AnnoProm returns the prometheus annotations for a pod.
-func AnnoProm(port int32, path string) map[string]string {
-	return map[string]string{
-		"prometheus.io/scrape": "true",
-		"prometheus.io/port":   fmt.Sprintf("%d", port),
-		"prometheus.io/path":   path,
-	}
-}
-
-// AnnoAdditionalProm returns the additional prometheus annotations for a pod.
-// Some pods may have multiple prometheus endpoints.
-// We assume the same path is used for all endpoints.
-func AnnoAdditionalProm(name string, port int32) map[string]string {
-	return map[string]string{
-		fmt.Sprintf("%s.prometheus.io/port", name): fmt.Sprintf("%d", port),
-	}
-}
