--- pkg/utils/k8s/annotation_test.go
@@ -1,35 +1,0 @@
-// Copyright 2024 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package k8s
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestPromAnno(t *testing.T) {
-	anno := AnnoProm(8234, "/metrics")
-	require.Equal(t, map[string]string{
-		"prometheus.io/scrape": "true",
-		"prometheus.io/port":   "8234",
-		"prometheus.io/path":   "/metrics",
-	}, anno)
-
-	annoAdditional := AnnoAdditionalProm("tiflash.proxy", 20292)
-	require.Equal(t, map[string]string{
-		"tiflash.proxy.prometheus.io/port": "20292",
-	}, annoAdditional)
-}
